@@ -1,25 +1,25 @@
 name: direct_dialer
 description: Allows Flutter applications to directly dial a phone number.
-version: 1.0.1+1
+version: 1.0.1
 homepage: https://github.com/GroovinChip/direct_dialer
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.12.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
     
-  android_intent_plus: ^1.0.1
-  connectivity_plus: ^1.0.1
-  device_info_plus: ^1.0.1
-  permission_handler: ^8.0.0+2
-  url_launcher: ^6.0.2
+  android_intent_plus: ^3.1.1
+  connectivity_plus: ^2.3.3
+  device_info_plus: ^3.2.4
+  permission_handler: ^9.2.0
+  url_launcher: ^6.1.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  lints: ^2.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
